FinalsMVP show finals stats instead of regular season stats on GodMode update
